Apparently I fixed the ram and os commands inadvertently. People are telling me they didn't work before and do now.. I guess that's just a lucky bonus addition.
Added a little thing at the top of docs/WHATSNEW that indicates
which order the lines therein consist of according to when the
changes they represent were implemented. Made this sound more
technical than it should for no reason at all.

More changes to command parsing involving 0 argument commands
being supplied with extra arguments. If the input type is
ADDRESS_INVOKE (the bot was called upon by name), the bot now
anyway, because it's going to output an "I don't know"
response from show_url if the topic doesn't exist. This doesn't
happen with DIRECT_INVOKE invocation because the text may have
been normal channel conversation that was mistaken as a command
by the bot.