Add ZADD alias for RANDOMSTUFF
Summary:
UnderNet #Win used `zadd` instead of `rs` as convenient alias
to add a random entry. It's even documented like this in the
former channel manual.

To be compatible with this manual, we add this alias.

Fixes T1150.

Test Plan:
Works for randomstuff:
21:24:30 < Dereckson> Gizmo: zadd foo bar
21:24:31 < Gizmo> Dereckson, done. There are now 83 randomstuffs.

Works for random replies for an info2 entry:
21:24:35 < Dereckson> Gizmo: zadd ~foo bar
21:24:35 < Gizmo> Dereckson, done. There are now 4 randomstuffs.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1150

Differential Revision: https://devcentral.nasqueron.org/D915